swarmguard/recovery-manager/requirements.txt
docker==7.1.0
PyYAML==6.0.2
Flask==3.1.0
requests==2.32.3
